+++ parsyfiles/tests/parsing_capabilities_by_type/test_parse_unions.py
@@ -1,6 +1,6 @@
 import os
 
-from typing import Union
+from typing import Union, Dict
 
 from parsyfiles import RootParser
 
@@ -13,6 +13,8 @@ def get_path(*args):
 
 
 def test_union_1(root_parser: RootParser):
+    """ Tests that parsing a Union works """
+
     class A:
         def __init__(self, foo: str):
             self.foo = foo
@@ -26,6 +28,8 @@ def test_union_1(root_parser: RootParser):
 
 
 def test_union_2(root_parser: RootParser):
+    """ Tests that parsing a collection of Union works """
+
     class A:
         def __init__(self, foo: str):
             self.foo = foo
@@ -38,3 +42,38 @@ def test_union_2(root_parser: RootParser):
     assert len(items) == 2
     assert type(items['a']) == A
     assert type(items['b']) == B
+
+
+def test_union_recursive_1(root_parser: RootParser):
+    """ Tests that you can parse infinitely-nested dictionaries from a folder using forward references """
+
+    class A:
+        def __init__(self, foo: str):
+            self.foo = foo
+
+    # First (preferred) way
+    InfiniteRecursiveDictOfA = Dict[str, Union[A, 'InfiniteRecursiveDictOfA']]
+
+    items = root_parser.parse_item(get_path('test2'), InfiniteRecursiveDictOfA)
+
+    assert type(items['a']['a']['a']) == A
+    assert type(items['a']['a']['b']) == A
+    assert type(items['a']['b']) == A
+    assert type(items['b']) == A
+
+    # Less preferred way, but check that it works too
+    InfiniteRecursiveDictOfA2 = Union[A, Dict[str, 'InfiniteRecursiveDictOfA2']]
+
+    items = root_parser.parse_collection(get_path('test2'), InfiniteRecursiveDictOfA2)
+
+    assert type(items['a']['a']['a']) == A
+    assert type(items['a']['a']['b']) == A
+    assert type(items['a']['b']) == A
+    assert type(items['b']) == A
+
+    # This is a forward reference that is equivalent to 'A'.
+    # It should be handled correctly by parsyfiles so as not to lead to infinite recursiong
+    InfiniteRecursiveDictOfA3 = Union[A, 'InfiniteRecursiveDictOfA3']
+
+    item = root_parser.parse_item(get_path('test2/b'), InfiniteRecursiveDictOfA3)
+    assert type(item) == A